refactor(repo C11a): extract vector search to engine/repository/vec.ts
Move searchVec, mapVectorMatchesToDocuments, getMainDatabasePath, and the
_vecUnavailableWarned module-level flag out of engine/repository.ts. Expose
getCollectionShardCount and getEmbedding as named exports so vec.ts can call
back into the parent module without duplication; later clusters relocate them.

Build: clean. Tests: 970/970 pass.
