+++ engine/repository/index.ts
@@ -16,4 +16,5 @@ export * from "./collections.js";
 export * from "./context-annotations.js";
 export * from "./content.js";
 export * from "./indexing.js";
+export * from "./vec.js";
 export type * from "./types.js";
